api Cabins update

+++ src/services/apiCabins.ts
@@ -2,6 +2,11 @@ import type { CabinType } from "../types/cabinType";
 import type { FormInputType } from "../types/formInputType";
 import { supabase, supabaseUrl } from "./supabase";
 
+interface EditCabinType{
+  newCabin: FormInputType,
+  id?: number,
+}
+
 export async function getCabins(): Promise<CabinType[]> {
   const { data, error } = await supabase.from("cabins").select("*");
   if (error) {
@@ -20,21 +25,36 @@ export async function deletCabins(id: number) {
   }
 }
 
-export async function createCabin(newCabin: FormInputType) {
-  console.log(newCabin.image[0].name)
+export async function createEditCabin({newCabin, id}: EditCabinType) {
+  const hasImagePath = typeof(newCabin.image) === "string"
   const imageName = `${Math.random()}-${newCabin.image[0].name}`.replace("/", "");
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath ? newCabin.image : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
   // 1. Create cabins
-  const { data, error } = await supabase.from("cabins").insert([{ ...newCabin, image: imagePath }]).select();
+  let query = supabase.from("cabins");
+
+  // A) CREATE
+  if (!id) {
+    query = query.insert([{ ...newCabin, image: imagePath }]);
+  }
+  // B) EDIT
+  if (id){
+    query = query.update({ ...newCabin, image: imagePath }).eq("id", id)
+  }
+  const { data, error } = await query.select().single();
+
   if (error) {
     console.error(error);
     throw new Error("Cabin could not be created");
   }
-  // 2. Upload image
-  const { error: storageError } = await supabase.storage.from("cabin-images").upload(imageName, newCabin.image[0]);
-  // 3. Delete the cabin if there was an error uploading the image
-  if (storageError) {
-    console.log(storageError)
-    await supabase.from("cabins").delete().eq("id", data.id);
+  if (!id || (id && !hasImagePath)){
+    // 2. Upload image
+    const { error: storageError } = await supabase.storage.from("cabin-images").upload(imageName, newCabin.image[0]);
+    
+    // 3. Delete the cabin if there was an error uploading the image
+    if (storageError) {
+      console.log(storageError);
+      console.log("here")
+      await supabase.from("cabins").delete().eq("id", data.id);
+    }
   }
 }
